# aspnet-server-1/src/IO.Swagger/Models/MongoLocation.cs
/*
 * Walking Skeleton API
 *
 * This is the API for the Walking Skeleton.
 *
 * OpenAPI spec version: 1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using MongoDB.Bson;
using MongoDB.Bson.Serialization.Attributes;

namespace IO.Swagger.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class MongoLocation
    {         
        [BsonId]
        public ObjectId ID {get; set;}
        /// <summary>
        /// Gets or Sets Id
        /// </summary>       
        public long? id { get; set; }

        /// <summary>
        /// Gets or Sets DeviceId
        /// </summary>       
        public long? deviceId { get; set; }

        /// <summary>
        /// Gets or Sets LocationValue
        /// </summary>
        public string locationValue { get; set; }      


        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Location {\n");
            sb.Append("  Id: ").Append(id).Append("\n");
            sb.Append("  DeviceId: ").Append(deviceId).Append("\n");
            sb.Append("  LocationValue: ").Append(locationValue).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }       
    }
}