# aspnet-server-1/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Walking Skeleton API",
    "description" : "This is the API for the Walking Skeleton.",
    "version" : "1"
  },
  "servers" : [ {
    "url" : "url1",
    "description" : "production server"
  }, {
    "url" : "url2",
    "description" : "testing server"
  } ],
  "paths" : {
    "/Locations" : {
      "get" : {
        "summary" : "returns all locations as a list",
        "operationId" : "getLocations",
        "responses" : {
          "200" : {
            "description" : "a JSON array of locations",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Location"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "404" : {
            "description" : "no locations are in the database"
          }
        }
      }
    },
    "/Location/{deviceId}" : {
      "get" : {
        "summary" : "returns the location associated with device ID",
        "operationId" : "getLocation",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "The ID of the device.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "a json object of a location",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Location"
                }
              }
            }
          },
          "400" : {
            "description" : "retrival of location went wrong"
          },
          "404" : {
            "description" : "no device with that id"
          }
        }
      },
      "delete" : {
        "summary" : "deletes a location",
        "operationId" : "deleteLocation",
        "parameters" : [ {
          "name" : "deviceId",
          "in" : "path",
          "description" : "The ID of the device.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "minimum" : 1,
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "location deleted successfully"
          },
          "400" : {
            "description" : "deletion of location went wrong"
          },
          "404" : {
            "description" : "no location associated with devive ID"
          }
        }
      }
    },
    "/Location" : {
      "post" : {
        "summary" : "adds a new location",
        "operationId" : "postLocation",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/Location"
        },
        "responses" : {
          "201" : {
            "description" : "location added successfully"
          },
          "400" : {
            "description" : "location added unsuccessfully"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Location" : {
        "required" : [ "deviceId", "id", "locationValue" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "deviceId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "locationValue" : {
            "type" : "string",
            "example" : "some RSSI value"
          }
        },
        "example" : {
          "id" : 0,
          "deviceId" : 6,
          "locationValue" : "some RSSI value"
        }
      }
    },
    "parameters" : {
      "deviceId" : {
        "name" : "deviceId",
        "in" : "path",
        "description" : "The ID of the device.",
        "required" : true,
        "style" : "simple",
        "explode" : false,
        "schema" : {
          "minimum" : 1,
          "type" : "integer",
          "format" : "int64"
        }
      }
    },
    "requestBodies" : {
      "Location" : {
        "description" : "a json object of a location",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Location"
            }
          }
        },
        "required" : true
      }
    }
  }
}